Fix message preview template
HTML preview of message was not shown due to condition depending on
non-existing attribute message.newsletter.message_template.html.

HTML version of message is supposed to be optional. With database
templates it depended on existence of HTML template but with current
file based templates default HTML template for message always exist. So
other mechanism is needed.