Use uuid version of activation emails. Support old activation format
The old email based format should work, but new urls are generated with the activation code as UUID, and no email.

There is one particular case where an update could be sent without an activation code or an email, which I’m really  not sure how that would have worked? This is on test_web.py test_update_request_activate_form. I’ve kind of disabled this, or changed what it was originally testing.