Feature: Add Attachment to a Message (#334)
This PR adds a new Attachment model which has a ForeignKey to Message. With this it is (optionally) possible to upload one or more files for a Message. When the message is send out the files are attached as normal e-mail attachments. The "read online" pages include a download link to these files.